watson-service/requirements.txt
asn1crypto==0.24.0
astroid==1.5.3
certifi==2017.11.5
cffi==1.11.2
chardet==3.0.4
click==6.7
cryptography==2.1.4
docutils==0.14
ecdsa==0.13
Flask==0.12.2
Flask-Cors==3.0.3
future==0.16.0
futures==3.1.1
ibm-cos-sdk==2.0.0
ibm-cos-sdk-core==2.0.0
ibm-cos-sdk-s3transfer==2.0.0
idna==2.6
isort==4.2.15
itsdangerous==0.24
Jinja2==2.10
jmespath==0.9.3
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
numpy==1.13.3
olefile==0.44
pandas==0.21.1
peewee==2.10.2
Pillow==4.3.0
psycopg2==2.7.3.2
pycparser==2.18
pycrypto==2.6.1
pylint==1.7.4
pyOpenSSL==17.5.0
pysolr==3.6.0
python-dateutil==2.6.1
python-jose==1.4.0
python-resize-image==1.1.11
pytz==2017.3
requests==2.18.4
requests-futures==0.9.7
scikit-learn==0.19.1
six==1.11.0
sphinx-rtd-theme==0.2.4
urllib3==1.22
watson-developer-cloud==1.0.2
Werkzeug==0.12.2
wrapt==1.10.11
xmltodict==0.11.0
